pyproject.toml
[tool.poetry]
name = "micro-services"
version = "0.1.0"
description = ""
authors = ["Sivan wolberg <sivan@wolberg.pro>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
flask = "^2.3.2"
Flask-Migrate = "^4.0.7"
Flask-SQLAlchemy = "^3.1.1"
Flask-WTF = "^1.2.1"
Flask-Bcrypt = "^1.0.1"
gunicorn = "^22.0.0"
email-validator = "^2.1.1"
Flask-Mail = "^0.10.0"
flask-rabbitmq = "^0.0.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
validator-collection = "^1.5.0"
SQLAlchemy-Utils = "^0.41.2"
marshmallow = "^3.21.3"
marshmallow-sqlalchemy = "^1.0.0"
Flask-Cors = "^4.0.1"
flask-restx = "^1.3.0"
descope = "^1.6.5"
requests = "^2.32.3"
cryptography = "^42.0.8"
charset-normalizer = "^3.3.2"
dynaconf = "^3.2.5"
Flask-JWT-Extended = "^4.6.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
black = {version = "^23.1a1", allow-prereleases = true}
mypy = "^0.991"

[tool.pylint.MASTER]
ignore-paths = [ "^server/migrations/.*$" ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint."MESSAGE CONTROL"]
disable = [
    "C0415", # Disable Import outside toplevel
]

[tool.black]
exclude = '''
  /migrations/
'''